requirements.txt
aws-cdk-lib==2.88.0
constructs>=10.0.0,<11.0.0
youtube_transcript_api==0.6.0
boto3==1.26.142